code smells & added component for form
