+++ platform/components/CustomFooter/index.tsx
@@ -3,11 +3,11 @@
 import Image from "next/image";
 import React from "react";
 
-import CustomButton from "@/components/CustomButton";
 import Link from "next/link";
 import { FooterSocials } from "@/constants";
+import Contact from "../CustomTextField/Contact";
 
-const Contact = () => {
+const CustomFooter = () => {
   return (
     <section className="flexCenter relative ">
       <Image
@@ -71,63 +71,7 @@ const Contact = () => {
                 </div>
               </div>
             </div>
-            <div>
-              <form
-                className="flex h-[624.04px] w-[514.56px] flex-col rounded-[10px] bg-white p-12"
-                action={() => {}}
-              >
-                <span className="text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
-                  Your Name
-                </span>
-                <input
-                  className="borderBlack mt-4 h-14 rounded-full border-2 border-gray-200 px-4"
-                  name="name"
-                  type="name"
-                  required
-                  maxLength={500}
-                  placeholder="Full name"
-                />
-                <span className="mt-8 text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
-                  INPUT FIELD
-                </span>
-                <input
-                  className="borderBlack mt-4 h-14 rounded-full border-2 border-gray-200 px-4"
-                  name="senderEmail"
-                  type="email"
-                  required
-                  maxLength={500}
-                  placeholder="name@mail.com"
-                />
-                <span className="mt-8 text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
-                  YOUR MESSAGE
-                </span>
-                <textarea
-                  className="my-3 mt-4 h-52 rounded-lg border border-black/10 p-4"
-                  name="message"
-                  required
-                  maxLength={500}
-                  placeholder="Message"
-                />
-                <div className="flexCenter mt-8 flex-1">
-                  <label htmlFor="c1">
-                    <div className="flexCenter group rounded-[10px]">
-                      <input
-                        id="c1"
-                        type="checkbox"
-                        className="h-6 w-6 cursor-pointer appearance-none rounded-[10px] bg-teal-400 text-teal-400 focus:ring-teal-100"
-                      />
-                    </div>
-                  </label>
-                  <p className="cursor-pointer pl-2 leading-relaxed text-slate-900 text-opacity-40">
-                    Send me a copy
-                  </p>
-                  <CustomButton
-                    className="h-[54.74px] w-[131.38px] bg-teal-400"
-                    text="Send"
-                  />
-                </div>
-              </form>
-            </div>
+            <Contact />
           </div>
           <div className="mb-10 mt-48 border bg-white bg-opacity-20" />
           <div className="mb-20 flex justify-between">
@@ -150,4 +94,4 @@ const Contact = () => {
   );
 };
 
-export default Contact;
+export default CustomFooter;
