+++ platform/components/CustomTextfield/Contact.tsx
@@ -1,38 +1,64 @@
-// import React from "react";
+import React from "react";
 
-// const Contact = () => {
-//   return (
-//     <section>
-//       <div>
-//         <div className="font-['DM Sans'] h-[56.93px] w-[448.87px] text-[42px] font-bold leading-[52px] text-white">
-//           Let’s Keep in Touch
-//         </div>
-//         <div className="font-['DM Sans'] h-[140.14px] w-[436.83px] text-[22px] font-medium leading-loose text-white">
-//           We have created a new product that will help designers, developers and
-//           companies create websites for their startups quickly and easily.
-//         </div>
-//         <div>
-//           {/* <Image src  */}
-//           <div className="font-['DM Sans'] h-[28.46px] w-[263.85px] text-base font-normal leading-relaxed text-white">
-//             +1 555 505 5050
-//           </div>
-//         </div>
-//         <div>
-//           {/* <Image src  */}
-//           <div className="font-['DM Sans'] h-[28.46px] w-[251.81px] text-base font-normal leading-relaxed text-white">
-//             info@designmodo.com
-//           </div>
-//         </div>
-//         <div>
-//           {/* <Image src  */}
-//           <div className="font-['DM Sans'] h-[85.39px] w-[275.89px] text-base font-normal leading-relaxed text-white">
-//             San Francisco, CA560 Bush St & 20th Ave, Apt5 San Francisco, 230909
-//           </div>
-//         </div>
-//       </div>
-//       <form></form>
-//     </section>
-//   );
-// };
+import CustomButton from "@/components/CustomButton";
 
-// export default Contact;
+const Contact = () => {
+  return (
+    <div>
+      <form className="flex h-[624.04px] w-[514.56px] flex-col rounded-[10px] bg-white p-12">
+        <span className="text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
+          Your Name
+        </span>
+        <input
+          className="borderBlack mt-4 h-14 rounded-full border-2 border-gray-200 px-4"
+          name="name"
+          type="name"
+          required
+          maxLength={500}
+          placeholder="Full name"
+        />
+        <span className="mt-8 text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
+          INPUT FIELD
+        </span>
+        <input
+          className="borderBlack mt-4 h-14 rounded-full border-2 border-gray-200 px-4"
+          name="senderEmail"
+          type="email"
+          required
+          maxLength={500}
+          placeholder="name@mail.com"
+        />
+        <span className="mt-8 text-sm font-bold uppercase leading-relaxed tracking-widest text-violet-950">
+          YOUR MESSAGE
+        </span>
+        <textarea
+          className="my-3 mt-4 h-52 rounded-lg border border-black/10 p-4"
+          name="message"
+          required
+          maxLength={500}
+          placeholder="Message"
+        />
+        <div className="flexCenter mt-8 flex-1">
+          <label htmlFor="c1">
+            <div className="flexCenter group rounded-[10px]">
+              <input
+                id="c1"
+                type="checkbox"
+                className="h-6 w-6 cursor-pointer appearance-none rounded-[10px] bg-teal-400 text-teal-400 focus:ring-teal-100"
+              />
+            </div>
+          </label>
+          <p className="cursor-pointer pl-2 leading-relaxed text-slate-900 text-opacity-40">
+            Send me a copy
+          </p>
+          <CustomButton
+            className="h-[54.74px] w-[131.38px] bg-teal-400"
+            text="Send"
+          />
+        </div>
+      </form>
+    </div>
+  );
+};
+
+export default Contact;
