+++ platform/components/CustomTextfield/index.tsx
@@ -5,7 +5,6 @@ import React from "react";
 import Hero from "./Hero";
 import Projects from "./Projects";
 import Crew from "./Crew";
-// import Contact from "./Contact";
 
 const CustomTextField = () => {
   return (
@@ -13,7 +12,6 @@ const CustomTextField = () => {
       <Hero />
       <Projects />
       <Crew />
-      {/* <Contact /> */}
     </>
   );
 };
